[Decomp][SpellMgr] Cohesion split: extract pet spell loaders to SpellMgrPet.cpp
