+++ src/game/Object/SpellMgr.cpp
@@ -2807,177 +2807,7 @@ SpellEntry const* SpellMgr::SelectAuraRankForLevel(SpellEntry const* spellInfo, 
 
 
 
-/**
- * @brief Loads spell pet aura mappings from the database.
- */
-void SpellMgr::LoadSpellPetAuras()
-{
-    mSpellPetAuraMap.clear();                               // need for reload case
 
-    uint32 count = 0;
-
-    //                                                0      1    2
-    QueryResult* result = WorldDatabase.Query("SELECT `spell`, `pet`, `aura` FROM `spell_pet_auras`");
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-        sLog.outString(">> Loaded %u spell pet auras", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 spell = fields[0].GetUInt32();
-        uint32 pet = fields[1].GetUInt32();
-        uint32 aura = fields[2].GetUInt32();
-
-        SpellPetAuraMap::iterator itr = mSpellPetAuraMap.find(spell);
-        if (itr != mSpellPetAuraMap.end())
-        {
-            itr->second.AddAura(pet, aura);
-        }
-        else
-        {
-            SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell);
-            if (!spellInfo)
-            {
-                sLog.outErrorDb("Spell %u listed in `spell_pet_auras` does not exist", spell);
-                continue;
-            }
-            int i = 0;
-            for (; i < MAX_EFFECT_INDEX; ++i)
-            {
-                if ((spellInfo->Effect[i] == SPELL_EFFECT_APPLY_AURA &&
-                     spellInfo->EffectApplyAuraName[i] == SPELL_AURA_DUMMY) ||
-                    spellInfo->Effect[i] == SPELL_EFFECT_DUMMY)
-                    {
-                        break;
-                    }
-            }
-
-            if (i == MAX_EFFECT_INDEX)
-            {
-                sLog.outError("Spell %u listed in `spell_pet_auras` does not have dummy aura or dummy effect", spell);
-                continue;
-            }
-
-            SpellEntry const* spellInfo2 = sSpellStore.LookupEntry(aura);
-            if (!spellInfo2)
-            {
-                sLog.outErrorDb("Aura %u listed in `spell_pet_auras` does not exist", aura);
-                continue;
-            }
-
-            PetAura pa(pet, aura, spellInfo->EffectImplicitTargetA[i] == TARGET_PET, spellInfo->CalculateSimpleValue(SpellEffectIndex(i)));
-            mSpellPetAuraMap[spell] = pa;
-        }
-
-        ++count;
-    }
-    while (result->NextRow());
-
-    delete result;
-
-    sLog.outString(">> Loaded %u spell pet auras", count);
-    sLog.outString();
-}
-
-/// Some checks for spells, to prevent adding deprecated/broken spells for trainers, spell book, etc
-bool SpellMgr::IsSpellValid(SpellEntry const* spellInfo, Player* pl, bool msg)
-{
-    // not exist
-    if (!spellInfo)
-    {
-        return false;
-    }
-
-    bool need_check_reagents = false;
-
-    // check effects
-    for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-    {
-        switch (spellInfo->Effect[i])
-        {
-            case SPELL_EFFECT_NONE:
-                continue;
-
-                // craft spell for crafting nonexistent item (break client recipes list show)
-            case SPELL_EFFECT_CREATE_ITEM:
-            {
-                if (!ObjectMgr::GetItemPrototype(spellInfo->EffectItemType[i]))
-                {
-                    if (msg)
-                    {
-                        if (pl)
-                        {
-                            ChatHandler(pl).PSendSysMessage("Craft spell %u create item (Entry: %u) but item does not exist in item_template.", spellInfo->Id, spellInfo->EffectItemType[i]);
-                        }
-                        else
-                        {
-                            sLog.outErrorDb("Craft spell %u create item (Entry: %u) but item does not exist in item_template.", spellInfo->Id, spellInfo->EffectItemType[i]);
-                        }
-                    }
-                    return false;
-                }
-
-                need_check_reagents = true;
-                break;
-            }
-            case SPELL_EFFECT_LEARN_SPELL:
-            {
-                SpellEntry const* spellInfo2 = sSpellStore.LookupEntry(spellInfo->EffectTriggerSpell[i]);
-                if (!IsSpellValid(spellInfo2, pl, msg))
-                {
-                    if (msg)
-                    {
-                        if (pl)
-                        {
-                            ChatHandler(pl).PSendSysMessage("Spell %u learn to broken spell %u, and then...", spellInfo->Id, spellInfo->EffectTriggerSpell[i]);
-                        }
-                        else
-                        {
-                            sLog.outErrorDb("Spell %u learn to invalid spell %u, and then...", spellInfo->Id, spellInfo->EffectTriggerSpell[i]);
-                        }
-                    }
-                    return false;
-                }
-                break;
-            }
-        }
-    }
-
-    if (need_check_reagents)
-    {
-        for (int j = 0; j < MAX_SPELL_REAGENTS; ++j)
-        {
-            if (spellInfo->Reagent[j] > 0 && !ObjectMgr::GetItemPrototype(spellInfo->Reagent[j]))
-            {
-                if (msg)
-                {
-                    if (pl)
-                    {
-                        ChatHandler(pl).PSendSysMessage("Craft spell %u requires reagent item (Entry: %u) but item does not exist in item_template.", spellInfo->Id, spellInfo->Reagent[j]);
-                    }
-                    else
-                    {
-                        sLog.outErrorDb("Craft spell %u requires reagent item (Entry: %u) but item does not exist in item_template.", spellInfo->Id, spellInfo->Reagent[j]);
-                    }
-                }
-                return false;
-            }
-        }
-    }
-
-    return true;
-}
